backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cachetools>=7.0.5",
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-qdrant>=1.1.0",
    "pdfplumber>=0.11.9",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
    "slowapi>=0.1.9",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.3",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
